+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 
 本项目版本遵守 [Semantic Versioning] 和 [PEP-440]。
 
+## [1.0.23] - 2020-06-07
+### Changed
+- http_util 默认会将 response json 转为 dict, 同时可以通过 check_key 和 check_value 两个参数来检查返回结果是否符合预期
+- shell_util 检查是否成功，由 success 属性修改为 success() 方法
+
 ## [1.0.22] - 2020-06-03
 ### Added
 - 增加 db 之 redis lib